Module Grpc_eio.Client

module Rpc : sig ... end

type response_handler = H2.Client_connection.response_handler

type do_request = ?flush_headers_immediately:bool -> ?trailers_handler:(H2.Headers.t -> unit) -> H2.Request.t -> response_handler:response_handler -> H2.Body.Writer.t

do_request is the type of a function that performs the request

val call : service:string -> rpc:string -> ?scheme:string -> handler:'a Rpc.handler -> do_request:do_request -> ?headers:H2.Headers.t -> unit -> ('a * Grpc.Status.t, H2.Status.t) Stdlib.result

call ~service ~rpc ~handler ~do_request () calls the rpc endpoint given by service and rpc using the do_request function. The handler is called when this request is set up to send and receive data.